Rough pass on radio/checkbox work
