+++ src/components/control-checkbox-set/control-checkbox-set.test.tsx
@@ -1,10 +1,22 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import { testCases } from './control-checkbox-set-test-cases';
+import ControlCheckboxSet from './control-checkbox-set';
 
 describe('ControlCheckbox', () => {
-  let testCase;
-  let wrapper;
+
+  /*
+testCases.basic = {
+  description: 'basic',
+  component: ControlCheckboxSet,
+  props: {
+    id: 'testinput',
+    options: [
+      { label: 'Option 1', value: 'Option value 1' },
+      { label: 'Option 2', value: 'Option value 2' }
+    ],
+    onChange: jest.fn()
+  }
+};
 
   describe(testCases.basic.description, () => {
     beforeEach(() => {
@@ -42,6 +54,19 @@ describe('ControlCheckbox', () => {
     });
   });
 
+testCases.disabled = {
+  description: 'disabled',
+  component: ControlCheckboxSet,
+  props: {
+    id: 'testinput',
+    options: [
+      { label: 'Option 1', value: 'Option value 1' },
+      { label: 'Option 2', value: 'Option value 2', disabled: true }
+    ],
+    onChange: jest.fn()
+  }
+};
+
   describe(testCases.disabled.description, () => {
     beforeEach(() => {
       testCase = testCases.disabled;
@@ -60,6 +85,26 @@ describe('ControlCheckbox', () => {
     });
   });
 
+testCases.allOptions = {
+  description: 'all options',
+  component: ControlCheckboxSet,
+  props: {
+    id: 'testinput',
+    options: [
+      { label: <span>Option 1</span>, value: 'Option value 1' },
+      { label: <span>Option 2</span>, value: 'Option value 2' },
+      { label: <span>Option 3</span>, value: 'Option value 3' }
+    ],
+    validationError: 'oh no!',
+    value: ['Option value 1', 'Option value 2'],
+    autoFocus: true,
+    optional: true,
+    themeCheckboxContainer: 'w-full',
+    themeCheckbox: 'checkbox--red mr6',
+    themeControlWrapper: 'bg-red-light',
+    onChange: jest.fn()
+  }
+};
   describe(testCases.allOptions.description, () => {
     beforeEach(() => {
       testCase = testCases.allOptions;
@@ -72,4 +117,5 @@ describe('ControlCheckbox', () => {
       expect(wrapper).toMatchSnapshot();
     });
   });
+*/
 });
