+++ src/components/control-checkbox-set/examples/control-checkbox-set-example-options.tsx
